+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Unreleased changes
 
+- Support relative and absolute block height as the block identifier in block queries.
+- Add field `protocol_version` to `BlockInfo` which is the protocol version of the queried block.
 - Extend enum `PendingUpdateEffect` with variants for protocol version 6.
 - Introduce `ChainParametersV2` struct for protocol version 6.
 - Introduce generic `gas_reward` in `RewardParametersSkeleton` for supporting different versions of GASRewards.
+- The SDK requires node version 5.4 or later.
 
 ## 2.4.0